footer assets moved to ./src/assets/LandingPage/Footer
